+++ proxyproto/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix crc32c TLV algorithm.
+
 ## 0.2.0
 
 - No significant changes since `0.1.1`.